+++ src/main/resources/db/migration/all/V1.21__create_table_unlock_script_p2raw_pub_key.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS unlock_script_p2raw_pub_key(
   tx_id BIGINT,
-  tx_index BIGINT,
+  tx_index INT,
   script_size INT,
   signature_id BIGINT,
     PRIMARY KEY(tx_id,tx_index),
