+++ src/main/resources/db/migration/all/V1.22__create_table_unlock_script_other.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS unlock_script_other(
   tx_id BIGINT,
-  tx_index BIGINT,
+  tx_index INT,
   script_size INT,
   script_id BIGINT,
     PRIMARY KEY(tx_id,tx_index),
